% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{Correction-class}
\alias{Correction-class}
\alias{show,Correction-method}
\title{Pressure correction model}
\usage{
\S4method{show}{Correction}(object)
}
\arguments{
\item{object}{the object to display.}
}
\description{
The pressure-correction component of a fitted \code{\link{Appac-class}}
object.
}
\section{Slots}{

\describe{
\item{\code{covariates}}{Names of the correction covariates (e.g. air pressure).}

\item{\code{coefficients}}{Fitted coefficient (\code{kappa}) per covariate.}

\item{\code{reference.values}}{Reference value (e.g. \code{P_ref}) per covariate.}

\item{\code{samples}}{Per-cylinder correction inputs.}

\item{\code{fit.data}}{Stored kappa-fit input, used by
\code{\link{plot_area_pressure_fit}}.}
}}

